Use separate exports in logger
