--- packages/logger/src/logger/index.ts
@@ -1,2 +1,0 @@
-export * from "./format.js";
-export * from "./winston.js";
