+++ packages/logger/src/loggers/empty.ts
