+++ packages/logger/src/utils/consoleTransport.ts
