+++ packages/logger/test/unit/logger/json.test.ts
@@ -3,7 +3,7 @@ import "../../setup.js";
 import {expect} from "chai";
 import {fromHexString, toHexString} from "@chainsafe/ssz";
 import {LodestarError} from "@lodestar/utils";
-import {logCtxToJson, logCtxToString} from "../../../src/logger/json.js";
+import {logCtxToJson, logCtxToString} from "../../../src/utils/json.js";
 
 describe("Json helper", () => {
   const circularReference = {};
