+++ packages/logger/test/unit/logger/timeFormat.test.ts
@@ -1,6 +1,6 @@
 import "../../setup.js";
 import {expect} from "chai";
-import {formatEpochSlotTime} from "../../../src/logger/timeFormat.js";
+import {formatEpochSlotTime} from "../../../src/utils/timeFormat.js";
 
 describe("logger / util / formatEpochSlotTime", () => {
   const nowSec = 1619171569;
