+++ packages/logger/test/unit/logger/winston.test.ts
@@ -3,7 +3,8 @@ import {expect} from "chai";
 import {MESSAGE} from "triple-beam";
 import Transport from "winston-transport";
 import {LodestarError, LogLevel} from "@lodestar/utils";
-import {LogData, LogFormat, logFormats, TimestampFormatCode, WinstonLoggerNode} from "../../../src/index.js";
+import {LogData, LogFormat, logFormats, TimestampFormatCode} from "../../../src/index.js";
+import {WinstonLoggerNode} from "../../../src/loggers/node.js";
 
 type WinstonLog = {[MESSAGE]: string};
